Add sanity tests on the json fields tables
If invalid data is in any of the tables it's an error and the program
will abort. These functions currently test for invalid field_type and
that includes making sure that the only element with JSON_NULL is the
one with the NULL name (which is supposed to be the last element). Other
tests could be devised and might be in time but this is a good start.

Sequenced exit codes.