jinfochk and jauthchk: additional checks added
Add empty_override and rule 2a/2b size mismatches: If empty_override ==
true and either the rule 2a or 2b size != 0 then it's an issue; if false
and either rule 2a or 2b size == 0 it's an issue. This has also been
added to mkiocccentry: if either size is 0 then warn about empty prog.c.

The 'test_mode' field is now common to both .info.json and .author.json
and we verify test_mode consistency with fnamchk: that is to say that if
test_mode == true the tarball filename must start with "entry.test-"; if
false it must be a UUID. This is done by adding two options to fnamchk:
-u: if filename is a test mode filename then it's an error; -t: if
filename is a UUID filename it is an error.

I've bumped the version of fnamchk, jinfochk, jauthchk and mkiocccentry
for the above changes.