Check for NULL elements in json fields tables
In other words if the name of a field that's not the last in the table
is NULL it's an error. This is done via calculating the size of the
table and iterating through the table as long as the end has not been
reached AND the current name != NULL.

This commit also fixes a bug where I had the thought of checking this
very thing yesterday but did not in fact add it: I inadvertently left
some elements after the NULL element (that's supposed to be last). This
didn't cause any problems but shouldn't be there anyway (and now that
this test is added it would cause jinfochk to fail with an error due to
an error in the table).

Sequenced exit codes.