jinfochk/jauthchk: initial extraction of arrays
The keyword is _extraction_: the parsing is NOT done. Instead the arrays
are extracted into a char * (and a strdup()d copy) for parsing as well
as moving past that point in the file. I am aware (and I noted it in the
comments) that if the array is last in the file there's an error but I
will worry about this later: I just wanted this committed since it's now
cleaned up.

I have added two files to test_JSON and I fixed one as well.

Also: if test mode is true (this does not mean the test mode in the json
files though perhaps that could be considered as well: I mean instead -t
passed to the util) then do not return 1 if issues were found. The
warnings will still be printed but the exit status will be 0.

Comment fixes, clean up and typos fixed.

Insignificant but in jinfochk.c and jauthchk.c: strtok_r(): savefield ->
saveptr.