Initial parsing of .info.json manifest
Currently what it does is it checks for each field name and verifies
that it's valid. For instance the field "prog_c" with value "program.c"
would trigger an issue. What it does not do is verify that every field
is there and it does not check that the extra data files are unique.
This will come at a future commit.

There's also some room for clean up but make test works and changing a
few things in the .info.json manifest array does trigger the issues I
would expect.