Move JSON_ARRAY case statements
In function check_found_info_json_fields() we don't actually need the
JSON_ARRAY_NUMBER or JSON_ARRAY_BOOL since the only value type in the
manifest array is string. It might also be that the
JSON_ARRAY_{STRING,NUMBER_BOOL} can actually be removed but I'll have to
look into that another time.

Also to point out in the last commit (I think) which I accidentally
committed json.c changes: instead of adding another copy of the same
value in the same field just increment the count of the field's value.
This was not supposed to be committed yet but it shouldn't cause a
problem either. It will possibly be used in a future commit.