jinfochk/jauthchk: call malloc_json_decode_str()
This is currently used on the values (for jinfochk each field's value as
well as the manifest array's field's values; for jauthchk only the
values of the normal fields as arrays are not yet parsed). The fields
common to both .info.json and .author.json also have this done.

I also added at least one call to not_reached() in a place I left it out
(possibly there were two places I noticed this).