jinfochk and jauthchk: check for empty strings
After skipping spaces etc., extracting the next field and skipping past
the array, if there's nothing left break out of the loop. When the
array_val_esc and val_esc are empty strings (I'm not even sure this is
possible but I'm doing it out of an abundance of caution) we also break
out of the loop.

I also removed a comment about a bug that was fixed yesterday.