jinfochk/jauthchk: Abort on invalid field in file
This solves the problem of e.g.:

    Warning: malloc_json_decode: non-strict encoding found \-escaped char: "
    FATAL[30]: check_author_json: malloc_json_decode_str(): invalidly formed field 'IOCCC_info_version' value '"1.0 2022-02-23"' or malloc failure in file test_work/test-0/.info.json

The problem was that if the field name is not expected we don't remove
the '"'s surrounding the name. But since an invalid field suggests an
invalid file this is not a problem.

This check can be safely done on jauthchk too despite the fact arrays
are not yet parsed and so the comments suggesting otherwise has been
removed.

I also changed the 'invalidly formed value ...' message to include the
field name as well to help identify problems.