New version release: bug fixes w/JSON data types
This commit fixes a number of problems with testing the validity of json
data types. The below is a list of the problems fixed (I'm not sure if I
have listed everything but these are what I can remember testing).

If a number or boolean has any '"' in it it's now flagged.

If a string does not have a leading '"' and trailing '"' and it's not a
null object it's now flagged.

If a field name does not have a leading and trailing ''" it's now
flagged.

For jinfochk the above checks are done for both array fields and normal
fields.

I've added a TODO in jinfochk.c noting that the some of the warnings
should be turned into errors after all testing is done. There are also
cases where not breaking out of the loop (or aborting) could cause an
infinite loop but these will be rooted out over time.