Bug fixes in commit c4c8d48037
There's now a new field in struct json_field: a bool 'can_be_empty': if
true it means the value of the field can be empty (in this case it means
the strings can be 'null' without '"'s). More cases of invalid values
for strings are tested for and more specific output is given as well.
These checks (which are done for array fields and normal fields in
jinfochk and normal fields for jauthchk - since arrays are not parsed
yet - and should probably be in its own function in json.c but for now
it's in three different parts).

In -t mode don't warn about missing fields: this is because many if not
all the files in test_JSON were made before some fields (e.g.
author_count) were made common so they're missing. I never use -t in
testing though so I'll see if there are any problems (and much work is
to be done on both tools).

On the subject of test_JSON files: since some of them have
timestamp_epoch as "Thu Jan  1 00:00:00 1970 UTC" instead of the format
"Thu Jan 01 00:00:00 1970 UTC" I have put the former in a new define:
ALT_TIMESTAMP_EPOCH and so the test in json.c now checks for both. This
variable is also exported to limit_ioccc.sh in the Makefile. When
everything is done this can be removed safely but much more has to be
done before this can happen.

There was another bug where reaching EOF would trigger an issue but this
has also been fixed.