Add sanity
...because sometimes we're all a little insane, some of us are a lot
insane and code is often very insane. :-)

What this commit does is add sanity.{h,c} which has a new function:
ioccc_sanity_chk() which runs sanity checks on all the tables: the
location table, the UTF-8 POSIX map and the three JSON fields tables.

The jinfochk and jauthchk fields tables checks were moved to json.c
because the tools which have a sanity check function all run the new
ioccc_sanity_chk (because if something is not right in one or more of
the tables it suggests there's a problem in the toolkit) and jinfochk.c
and jauthchk.c obviously have main() each as well as verbosity_level.