Don't complain about missing optional json fields
For example the extra_file field is not required in the .info.json
manifest. The structure was set up to support this but I neglected to
add this condition in the check. This is done for both the common fields
and the .info.json fields: it's not done for .author.json as that check
isn't in yet as not everything is parsed there.