Changes to make full use of posix_plus_safe()
Changed code to use posix_plus_safe() as needed.

Added a number of comments of the form where the choice
of case has not been decided.

    /* XXX - should the lower_only (2nd) arg to posix_plus_safe() be true or false? */

Adjusted comments and remarks in several places to refer to a
posix_plus_safe() regular expression.  We have have missed a few cases
where the comment does not refer to a regular expression, BTW.

Sequence exit codes.