Fix json-test.sh so it runs
This includes adding files to
test_JSON/author.json/{good,strict-good,strict-bad}: since the
directories must exist for the script to continue. The files added might
not be strictly :) correct: that is to say the strict-bad file is
all.random.text which isn't really the right test and the strict-good
test is the default .author.json from mkiocccentry which might not have
any strictness to check (the good file is also this file both called
default-author.json which is the default from `make test').

Typo fixes also made in json-test.sh.

In a future commit I will change the warn() calls to dbg() with DBG_NONE
so that all that has to be done when testing is over (and very much has
to be done still hence why make test does not run this script yet!) is
change the debug level. This means that the test script will still be
very chatty but that's necessary for now.