Add find_matching_quote() function in util.c
This function finds the matching '"' that's unescaped. This is currently
not used but very possibly will be in the future: it was necessary as
part of another function I was playing with but for now that's not in
(it might be useful but I've not determined this yet so it's not in):

    /* find_matching_quote	find the next unescaped '"'
     *
     * Assuming *q == '"' find the matching (closing) '"' (i.e. the next unescaped
     * '"') and return a pointer to it (or NULL if not found). If *q != '"' && *q !=
     * '\0' it will return 'q'; else it'll return q updated to either the end of the
     * string (which means NULL return value) or the matching unescaped '"'.
     *
     * NOTE: This function does not return on NULL pointer passed in but it can
     * return a NULL pointer: this happens if no matching '"' is found (or the
     * string was empty in the first place).
     */
    char *
    find_matching_quote(char *q)
    {
	/* ... */
    }

Sequenced exit codes.

Changed in json.c the string "un-escaped" to "unescaped": this is purely
a cosmetic change; unescaped is not actually a recognised word in the
dictionary (at least OED) but it feels more nature to have it
"unescaped" rather than "un-escaped". It's also more consistent with
some comments I have used before in the code (comments and/or code).