Finish jinfochk
According to all the test cases in test_JSON/info.json jinfochk is
completed. If more test cases are created these will be fixed as well.

Removed the exit code of jinfochk and jauthchk being dependent on -t
option. I'm not even sure what my original intent of this was but it's
not needed: -t is needed for the test_JSON files because some fields
were not common at the time of creation of the files and version changes
cause a problem too.

What's left for jinfochk is to change all the warn() calls on invalid
files to be instead dbg() calls but this can come later.

Since Landon added my name to jinfochk.h (and thank you again - it means
a great deal to me Landon! Truly) I have added it to jinfochk.c as well
now that it's finished.