Bugs and typos fixed
In find_json_field_in_table() fix potential NULL pointer dereference
(this wouldn't happen in the current code but if one were to retrieve
the number of elements in a table one of the ways to do that would have
run strlen() on a NULL pointer).

Use TBLLEN() macro for the SIZEOF_ JSON fields tables as well as for the
location table.

Typo fixes in utf8_posix_map.c.