Various bug fixes and updates
The common json field author_count is no longer common but instead only
in .author.json.

The common json field iocccsize_version is no longer common but instead
only in .info.json.

In the functions check_found_info_json_fields() and
check_common_json_fields() no longer allow missing fields to be not a
problem if test mode is enabled.  This was causing some tests to fail
and since the problematic fields are no longer common this should be
okay. Because not all fields are checked in jauthchk this was not done
so this is not a problem. The test option does still allow for different
versions of the tools to be there however.

This commit resolves a number of cases in the test_JSON directories
specific to jinfochk. As jauthchk does not yet parse the authors array
I'm not sure where the test files stand with it. I am aware of some
test_JSON files specific to jinfochk that are still a problem which will
be worked out in a future commit. There seems to be an issue with the
test script which has caused some confusion and this will have to be
worked out as well.