All test_JSON/info.json/good files pass all tests
Remove author_count from test_JSON/info.json/good files (that still had
it) as it's no longer a common field (as of yesterday it belongs only in
.author.json). This means that all the test_JSON/info.json/good files
now pass. The test_JSON/info.json/strict-good files already all passed.

There are a few invalid results of the test_JSON/info.json/bad (but not
test_JSON/info.json/strict-bad) which I will be looking into soonish
(probably next) but it appears that once the tests of these seven files
are resolved the only thing that will have to be done (until such a time
as additional cases come up) is to deal with the ','s in strings.