jinfochk/jauthchk: Don't abort on invalid fields
This is for testing purposes: previously it was an error and if say a
field that was not supposed to be in the file is (e.g. due to change in
common fields like what happened yesterday: some info.json files in bad/
also had author_count) in the file it would prevent the proper tests
from being run at the end because parsing would have been unfinished.
This is safe to do because we only dereference the pointers if != NULL
and additionally there still will be an issue triggered if there's an
invalid field.

Note that the issues variable is incremented in parsing but this is
different from in the checker: in the parser it means there's a field in
the file that should not be there but in the checker it means that the
field is simply not handled (the loop does check that the field is
actually valid and if it's not it aborts because it suggests a serious
problem as the field should never have been added to the list in the
first place). Thus in the checker we only warn about an unhandled field
but we do not increment the number of issues as it doesn't necessarily
mean the file is invalid: it simply means a field was not checked in the
checker function.

These changes are for both jinfochk and jauthchk.