Fix off by one error in posix_plus_safe()
The loop did not check the final character because the for loop was:

     for (i=start; i < len-1; ++i) {

Where len was strlen(str). The loop is now correctly:

     for (i=start; i < len; ++i) {

This does fix one issue in test_JSON/info.json/bad/info.title-2.json.
However there still is another problem in that file which is not yet
picked up: that will be addressed in a future commit.