Remove need for ALT_TIMESTAMP_EPOCH
Because many of the test_JSON files had the wrong TIMESTAMP_EPOCH tests
were failing so when I added the check for this field I added
ALT_TIMESTAMP_EPOCH which was defined as:

    #define ALT_TIMESTAMP_EPOCH "Thu Jan  1 00:00:00 1970 UTC"	/* gettimeofday epoch */

Then in the check function it checked for both ALT_TIMESTAMP_EPOCH and
TIMESTAMP_EPOCH with an XXX note that said once all testing is complete
to remove it. Instead to make the invalid epochs be correct I did:

    for i in */*; do sed -i 's/Thu Jan  1 00:00:00 1970 UTC/Thu Jan 01 00:00:00 1970 UTC/g' "${i}";done

Now Makefile, limit_ioccc.h and json.c have all been updated.  This
means that limit_ioccc.sh will no longer have the ALT_TIMESTAMP_EPOCH
but if necessary it could be added back by adding to limit_ioccc.h the
define as above and reverting the Makefile diff:

    -	${GREP} -hE '^#define (.*_VERSION|TIMESTAMP_EPOCH|ALT_TIMESTAMP_EPOCH|IOCCC_TOOLSET_RELEASE)' version.h limit_ioccc.h | \
    +	${GREP} -hE '^#define (.*_VERSION|TIMESTAMP_EPOCH|IOCCC_TOOLSET_RELEASE)' version.h limit_ioccc.h | \

Maybe only json.c should have been updated? I don't know but in any
event the files are now fixed to have the proper timestamp epoch.