Resolve several invalid json test results
This commit resolves the invalid test results for the following files:

- test_JSON/info.json/bad/info.tarball-0.json
- test_JSON/info.json/bad/info.contest_id-0.json
- test_JSON/info.json/bad/info.title-2.json
- test_JSON/info.json/bad/info.tarball-2.json

This leaves only two tests not succeeding which are in the files:
test_JSON/info.json/bad/info.manifest-25.json and
test_JSON/info.json/bad/info.manifest-24.json. This along with not
allowing for commas in strings is all that has to be done for jinfochk
(so far). The parser will have to change for this at a future date.