Added jcodechk.sh use top json-test.sh
After json-test.sh tests foo.json, it will look for foo.json.code.
If foo.json.code exists, it will run json-test.sh and compare
the codes generated by the test_prog with those found in foo.json.code.
If there is a mismatch, a warning is raised and the json-test.sh will
exit non-0 at the end.  Mismatch information will be appended to the logfile.

Output being appended to the logfile in json-test.sh is now also written to stderr.