Check NULL pointer value in json_field lists
Although a NULL pointer value should never be added it still should be
tested and previously only the strlen() was tested.

Also updated top of file comments in jinfochk.{h,c} and jauthchk.{h,c}
noting the status of the tools and that it's been decided that flex(1)
and bison(1) should be used and that this will take some time and
thought. Also mention that a new tool jparse will likely be written.

Sequenced exit codes.