Updates to README.md
Made certain things stand out more (Anthony Howe's name in bold instead
of italic, his GitHub handle in italic instead of no style), added more
information about jinfochk and jauthchk and added my name under txzchk
as well. I hadn't done it before as it didn't feel right but since it
was added to txzchk.c and txzchk.h by Landon after I finished it I
decided to finally add it here.

Note in the previous commit I changed it so that jinfochk and jauthchk
state that it's being worked on by me - not written by - as I don't
think it's right to say it's been written when it's incomplete though I
appreciate still the gesture very much and it's very very much an honour
and privilege to be helping out with these tools!