Fix commits da5e76f and 4d46d6e
When I changed tool set to toolkit I forgot to update the Makefile so
that the limit_ioccc.sh script has the right variable.

It also said in many files 'toolkit chain' and I don't think that's
correct so I removed the chain part. This was in header files, source
files and man pages.