Bug fixes in warn()/err() functions
The functions did not actually report the error when writing the
warning/error but rather they just printed the return value of the
preceding call. Now they use errno and strerror(errno). This was carried
over in jwarn() as well.

Also some minor typo fixes: specifically in the warning of output
errors:

    -	warn(__func__, "\nin err(%d, %s, %s ...): fflush returned error: %d\n", exitcode, name, fmt, ret);
    +	warn(__func__, "\nin err(%d, %s, %s, ...): fflush returned error: %s\n", exitcode, name, fmt, strerror(errno));

In other words a comma was added after the last '%s' before the '...'. I
already had done this for jwarn().