Add json.codes document + more calls to jwarn()
The json.codes document has a list of current json warnings issued by
the jwarn() function with a lot of helpful comments (like explaining the
reserved codes as they're currently used, the format of the warnings and
other things) and tips (like how to list any hints in the file as well
as how to get a list of codes and any associated notes based on the last
time the file was updated). Obviously this file will have to be updated
as I add more warnings to jwarn() which means that it's possible it's
out of date but I will try my best to add these (and update others if
necessary) when they're added. As of 14 March 2022 (today) I believe I
have all the few jwarn() warnings in the file.

Added a few calls to jwarn() in jinfochk.c and jauthchk.c. As well
changed the jwarn() call in json.c to be more like how it is in
jinfochk.c and jauthchk.c (the warning being a field found more times
than it is allowed which is an example where the code is the same across
both tools).