get_*_json_field()->add_*() + new error condition
Since the functions get_author_json_field(), get_info_json_field() and
get_common_json_field() add the field to the found list if they belong
to the respective list I have changed the prefix to add_ instead of
get_ so that I can do e.g. if (add_info_json_field(...)) which makes the
intent more obvious.

When jinfochk cannot add the 'manifest' field to the
found_info_json_fields list it is an error instead of a warning as this
indicates a real problem since the field MUST be in the info_json_fields
table which means that field cannot be NULL in add_info_json_field()
which means that add_found_info_json_field() MUST be called and if any
of these do not happen there is an error (if the function
new_json_field() or any other function it relies on cannot allocate
memory for the field it aborts so at that point if it didn't return this
wouldn't be an issue anyway) and the program should not continue.

Sequenced exit codes.