Tables check functions error message improvements
The error messages now state to fix the table and in which file and to
then recompile. This makes it just a bit more obvious where the problem
is.

Also move check for 'manifest' field into the function
check_json_fields_tables() so it's no longer needed in jinfochk.c
(though it still exists just in case).

Sequenced exit codes.