Add more JSON error/warning functions
Rename json.codes to json_warn.codes and add json_err.codes.

Add jerr() and jerrp() which are (mostly) analogous to err() and errp()
(error and error with errno functions). The difference is that they
have the same information that the jwarn() and the new jwarnp() (jwarn()
with errno information just like warnp()) function(s) have. The
functions look like (yes today is the 15th but the jwarnp() is almost
the same as the jwarn() which was added on the 13th):

/*
 * jwarnp - issue a JSON warning message with errno information
 *
 * given:
 *      code        warning code
 *      program     name of program e.g. jinfochk, jauthchk etc.
 *      name        name of function issuing the warning
 *      filename    filename of json file, "stdin" or NULL
 *      line        JSON line
 *      line_num    the offending line number in the json file
 *      fmt         format of the warning
 *      ...         optional format args
 *
 * Example:
 *
 *      jwarn(JSON_CODE(1), program, __func__, file, line, __LINE__, "unexpected foobar: %d", value);
 *
 * XXX As of 13 March 2022 the line will be empty but in time this should be
 * changed to be the offending JSON text and the offending JSON line number
 * (which currently is done but the way it's done might change). Instead the
 * field and values are usually shown as part of the optional format args.
 *
 * NOTE: We warn with extra newlines to help internal fault messages stand out.
 *       Normally one should NOT include newlines in warn messages.
 *
 * This function does not return on code < JSON_CODE_RESERVED_MIN (0).
 */
void
jwarnp(int code, char const *program, char const *name, char const *filename, char const *line, int line_num, char const *fmt, ...)
{
}

/*
 * jerr - issue a fatal JSON error message and exit
 *
 * given:
 *      exitcode        value to exit with
 *      program         program or NULL (will be set to "jchk")
 *      name            name of function issuing the error
 *      filename        file with the problem
 *      line            line with the problem (or NULL)
 *      line_num        line number with the problem or -1
 *      fmt             format of the warning
 *      ...             optional format args
 *
 * Example:
 *
 *      jerr(JSON_CODE(1), __func__, program, file, line, line_num, "bad foobar: %s", message);
 *
 * NOTE: We warn with extra newlines to help internal fault messages stand out.
 *       Normally one should NOT include newlines in warn messages.
 *
 * This function does not return.
 */
void
jerr(int exitcode, char const *program, char const *name, char const *filename, char const *line, int line_num, char const *fmt, ...)
{
}

/*
 * jerrp - issue a fatal error message with errno information and exit
 *
 * given:
 *      exitcode        value to exit with
 *      program         program or NULL (will be set to "jchk")
 *      name            name of function issuing the warning
 *      filename        file with the problem
 *      line            line with the problem or NULL
 *      line_num        line number of the problem or -1
 *      fmt             format of the warning
 *      ...             optional format args
 *
 * Example:
 *
 *      jerrp(JSON_CODE(1), program, __func__, file, line, line_num, "bad foobar: %s", message);
 *
 * NOTE: We warn with extra newlines to help internal fault messages stand out.
 *       Normally one should NOT include newlines in warn messages.
 *
 * This function does not return.
 */
void
jerrp(int exitcode, char const *program, char const *name, char const *filename, char const *line, int line_num, char const *fmt, ...)
{
}

Currently only four JSON error codes are defined (see json_err.codes)
and they're all used in three different functions as they have the same
purpose. There are other json errors that will share codes in a future
commit on another date.

The rationale behind these functions and the selected error codes is
that just like with warnings json errors are special in a way. Yes this
means that the exit codes can be the same as others with err()/errp()
but instead of showing FATAL[%d] it's JSON[%04d] to indicate it's a JSON
error. If I didn't note it elsewhere for the jerr() calls made right now
(codes 1 - 4) the file is actually json.c and the line number is
actually the line number of json.c (not the line that the table is on
but the check functions). This is because I couldn't think of a better
way to indicate that the line is actually not in a json file. This is a
work in progress!

Removed obsolete comment in dbg.c.