JSON warning codes and error codes updates
Added several new errors with reserved error codes. These error messages
are the same across json.c, jinfochk.c and jauthchk.c except for:
__FILE__ obviously depends on the file, __LINE__ depends on the line in
the file, __func__ depends on the function (though in one case it's not
the calling function which is the actual offender) and then the name of
the table in question depends on the table being checked/operated on. So
now reserved json error codes are (at least for now) those that are used
to verify that the (author|common|info)_json_fields tables are sane and
errors that are more or less the same in each tool.

The warnings for parsing json files (exception of invalid field but this
too might change later) that were json specific now just use warn():
this is because once the json parser is added these warnings will
probably not even exist so I'd rather save the warning codes for
something else.

Updated json_(err|warn).codes for the above changes.

Sequenced exit codes.

NOTE: I've decided to not use the jwarn()/jwarnp() or jerr()/jerrp()
functions for warnings/errors that might not exist when the parser is
added so not all warnings and errors will be added prior to the parser
though of course that isn't necessarily carved in stone.