Add initial jparse tool
This tool will in time parse blocks of JSON from the command line (as a
string or a file) or stdin (which can also be done by passing '-' as an
argument to the program). However right now it only attempts to read the
file (stdin or otherwise) or string and then if there's no embedded NUL
bytes it prints it out. Then it moves to the next arg until either an
error occurs or all have been processed.

The tool takes certain precautions so that stdin is never closed but
that it can also be used more than once (clearerr() is used on it in
order to clear the EOF and error flags). It also takes care that if
opening a file fails (fopen() returns NULL) it only aborts iff (if and
ONLY if!) there's an actual error; if the file does not actually exist
(errno == ENOENT) then it processes it as if it's not a file (since it's
not a file).

Updated .gitignore.

Sequenced exit codes.

If the first paragraph did not make it clear this tool is unfinished
because the json parser does not exist yet. A man page should be written
for this but I'll worry about that later.