Add more JSON_CODE_RESERVED codes (10 and 11)
The codes are as follows:

    # program: jchk check_found_(author|common|info)_json_fields
    JSON[0010]: (jauthchk|json|jinfochk).c: %d: illegal field name '%s' in found_(author|common|info)_json_fields list

    # program: jchk check_found_(author|common|info)_json_fields
    JSON[0011]: (jauthchk|json|jinfochk).c: %d: NULL pointer val for field '%s'

I had the thought of making all firewall checks that check for NULL
pointers use JSON_CODE_RESERVED(0) but I'm not sure if I want to do
that. I also had the thought of writing a script that verifies that
codes are not reused in a way they should not be but I'm not worried
about that right now yet either. There probably won't be many more
updates on the calls to jerr()/jerrp() or jwarn()/jwarnp() until after
the parser is completed but I'm not certain of this either.

Sequenced exit codes.