jparse: use ioccc_sanity_chks() + comment about -s
Call ioccc_sanity_chks() function to make sure everything is sane. This
is not so useful yet but I meant to have it originally and it's better
to make sure everything in the repo is sane.

The -s option is not currently used because json parsing is not yet
done. But it is set when -s is passed to the program as I wanted the
option there for future use. This is why the compiler flags for jparse
include -Wno-unused-but-set-variable.