Fix reading of file in jparse.c
My clever use of ENOENT to decide if it's an error in opening or if the
file simply does not exist (and so read it as if it was a string) was in
error because not only is there the function is_file() (that totally
slipped my mind) but I didn't think of abnormal files like directories.
This commit fixes that.