JSON warn and error codes / messages updates
The functions that add a field/value now take a 'char const *' called
json_filename for reporting the name of the offending file (some
already had this: in that case it was renamed to be 'json_filename'
usually from 'file' or 'filename'). At least for now the file is the
source file (the one that the function is in) and the messages will
refer to the offending file instead.

The json_{err,warn}.codes files have been updated to reflect the message
changes (it's possible I missed some though).

Some typo fixes and other minor changes in jparse.c. I just became aware
of another problem that also is in jstrdecode.c and/or jstrencode.c and
probably other tools as well which I'll look at next.