Use errp() instead of err() in jparse.c
There were calls to err() when errno was available so these have been
changed to refer to errp().

The issue I noted in the previous commit that stated that I became aware
of a problem in jstrencode/jstrdecode is this issue but it seems it was
actually already okay. There's a minor issue to do with the printing of
the IOCCC toolkit release which will come in the next commit (more tools
probably have to have this done). This was done for jparse.c in a
previous commit.