Fix call to warnp() of printing IOCCC toolkit
The diff looks like

	-	warnp(__func__, "printf error printing IOCCC toolkit release repository tag");
	+	warnp(__func__, "printf error printing IOCCC toolkit release repository tag: %s", IOCCC_TOOLKIT_RELEASE);

This makes it match the others where it shows the version the warning
string. There are certainly other tools where this has to be done
because when I added it it to all tools (aside from mkiocccentry) I just
yanked/pasted so the same code was copied to all files.