Make -S for strict and -s for string arg
The logic behind this is that string arg will be more commonly used and
more useful than strict mode and therefore although currently only one
tool has string arg (jparse) it feels more natural to have -s for string
and -S for strict mode. Discussing this with Landon we both had come to
this same though so the tools that have strict mode have been updated.
The two letters have been swapped in jparse.

Additionally the man pages for jinfochk and jauthchk were updated to
include this information (they did not yet have strict mode added to the
SYNOPSIS and OPTIONS sections).