Initial json flex/bison grammar added
This is very incomplete and there are certainly some things that are
wrong! As I learn more about flex(1)/bison(1) and also get into actually
working on the parsing things will be fixed. I _believe_ I have the case
of terminals/non-terminals correct but even that might be not entirely
the case.

I've not gone through the entire json grammar so both json_parser.l and
json_parser.y are incomplete in that way too (it's also incomplete in
that additional C code has to be written). It's also possible I have
some of the regular expressions wrong and the actions are not complete:
all they do is print what was found and then return the token type
(which might also be incomplete).

I've read only the very beginning of the book Flex and Bison and much
more has to be done there but I wanted these in. This commit involves
some other changes.

First of all the following #defines in json.h have been changed:

    JSON_STRING		-> JSON_CHARS
    JSON_ARRAY_STRING	-> JSON_ARRAY_CHARS
    JSON_NUMBER		-> JSON_NUM
    JSON_NULL		-> JSON_EOT /* EOT = end of table */

The reason is that these are also used in the parser and json.h will
eventually be #included (actually it already is but not yet used as
such).

The Makefile has been updated to run flex and bison on json_parser.l and
json_parser.y respectively but not by default: one has to do `make
json_parser'. This will also compile in the generated code which also
happens to be in the repo so it's probably better that only I do that
for now. There is a slight issue here: the rule has to be split into
two, one to generate the code and the other to compile. This is not a
problem itself but the first one won't work on systems without flex and
bison installed so it cannot be part of the default. Instead whenever I
update the .l/.y files I have to run `make json_parser` and recommit the
files. Eventually I plan on cleaning the output and then the need for
running flex and bison won't be there any more but this is quite a bit
down the road.

.gitignore has also been updated with 'json_parser' added to the list to
ignore.

On the subject of json_parser: right now (if it's not obvious) it is
entirely useless as it doesn't do anything at all. The importance of
this commit is the starting of the grammar and getting the repo in a
state that I can start updating it and working on the actual parser.
Then integrating the generated parser into the tools e.g. jparse,
jinfochk and jauthchk plus any others that might need the parser. I'm
wondering if maybe json_parser should become jparse and so the code in
jparse.c is added to the json_parser (which will be named jparse
instead) so that once the parser is done it will be right there to use.
This however is something I can work out later.